Presentation of the Social Work Video Game Demo
Social Work Game Simulation Demo

Ngorosha, M. & Vanmeter, J. (2020). Social Work Game Simulation Demo. Brescia University.

The Social Work Simulation is a demo of a video game played from a first-person point of view that explores the experience of a social worker and characters in the game. The game allows the player to face similar situations that a social worker would encounter, along with allowing them to empathize with characters. We are building the game using Unity, a game engine that allows creators to create 2D and 3D VR & AR visualizations for games, transportation, film, and animation among others. The purpose of our game is to help prepare future social workers for common scenarios they may face in the field. It gives players the option to make mistakes and correct them, in order to learn from their experiences. The first scenario in the game takes place in a kitchen. As we continue to develop the game with more scenarios, we plan to create an entire home in which different scenes happen.